+++ examples/celery_app.rs
@@ -31,7 +31,7 @@ async fn buggy_task() -> TaskResult<()> {
 #[celery::task(max_retries = 2)]
 async fn long_running_task(secs: Option<u64>) {
     let secs = secs.unwrap_or(10);
-    time::delay_for(Duration::from_secs(secs)).await;
+    time::sleep(Duration::from_secs(secs)).await;
 }
 
 // Demonstrates a task that is bound to the task instance, i.e. runs as an instance method.
@@ -59,9 +59,11 @@ enum CeleryOpt {
 #[tokio::main]
 async fn main() -> Result<()> {
     env_logger::Builder::from_env(Env::default().default_filter_or("info")).init();
+
     let opt = CeleryOpt::from_args();
+
     let my_app = celery::app!(
-        broker = AMQP { std::env::var("AMQP_ADDR").unwrap_or_else(|_| "amqp://127.0.0.1:5672/my_vhost".into()) },
+        broker = AMQPBroker { std::env::var("AMQP_ADDR").unwrap_or_else(|_| "amqp://127.0.0.1:5672/my_vhost".into()) },
         tasks = [
             add,
             buggy_task,
@@ -76,7 +78,7 @@ async fn main() -> Result<()> {
         ],
         prefetch_count = 2,
         heartbeat = Some(10),
-    );
+    ).await?;
 
     match opt {
         CeleryOpt::Consume => {