+++ src/app/tests.rs
@@ -4,17 +4,8 @@ use crate::protocol::MessageContentType;
 use crate::task::{Request, Signature, Task, TaskOptions, TaskResult};
 use async_trait::async_trait;
 use chrono::{DateTime, Duration, Utc};
-use futures::executor::block_on;
-use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
 use std::time::SystemTime;
-
-// Basic app with all defaults.
-static MOCK_BASIC_APP: Lazy<Celery<MockBroker>> = Lazy::new(|| block_on(build_basic_app()));
-
-// App with some custom configured options.
-static MOCK_CONFIGURED_APP: Lazy<Celery<MockBroker>> =
-    Lazy::new(|| block_on(build_configured_app()));
 
 async fn build_basic_app() -> Celery<MockBroker> {
     let celery = Celery::<MockBroker>::builder("mock-app", "mock://localhost:8000")
@@ -136,133 +127,136 @@ struct MultiplyParams {
     y: i32,
 }
 
-#[test]
-fn test_app_name() {
-    assert!(MOCK_BASIC_APP.name == "mock-app");
+#[tokio::test]
+async fn test_app_name() {
+    let app = build_basic_app().await;
+    assert!(app.name == "mock-app");
 }
 
 #[tokio::test]
 async fn test_add_task_registered() {
-    assert!(MOCK_BASIC_APP
-        .task_trace_builders
-        .read()
-        .await
-        .contains_key("add"));
+    let app = build_basic_app().await;
+    assert!(app.task_trace_builders.read().await.contains_key("add"));
 }
 
 #[tokio::test]
 async fn test_send_task() {
-    let result = MOCK_BASIC_APP.send_task(AddTask::new(1, 2)).await.unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
-    assert!(&message.headers.task == "add");
+    let app = build_basic_app().await;
+    let result = app.send_task(AddTask::new(1, 2)).await.unwrap();
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
+    assert!(message.headers.task == "add");
 }
 
 #[tokio::test]
 async fn test_send_task_with_countdown() {
-    let result = MOCK_BASIC_APP
+    let app = build_basic_app().await;
+    let result = app
         .send_task(AddTask::new(1, 2).with_countdown(2))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
-    assert!(&message.headers.eta.is_some());
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
+    assert!(message.headers.eta.is_some());
 }
 
 #[tokio::test]
 async fn test_send_task_with_eta() {
-    let result = MOCK_BASIC_APP
+    let app = build_basic_app().await;
+    let result = app
         .send_task(AddTask::new(1, 2).with_eta(DateTime::<Utc>::from(SystemTime::now())))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
-    assert!(&message.headers.eta.is_some());
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
+    assert!(message.headers.eta.is_some());
 }
 
 #[tokio::test]
 async fn test_send_task_with_expires_in() {
-    let result = MOCK_BASIC_APP
+    let app = build_basic_app().await;
+    let result = app
         .send_task(AddTask::new(1, 2).with_expires_in(10))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
-    assert!(&message.headers.expires.is_some());
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
+    assert!(message.headers.expires.is_some());
 }
 
 #[tokio::test]
 async fn test_send_task_with_expires() {
+    let app = build_basic_app().await;
     let dt = DateTime::<Utc>::from(SystemTime::now()) + Duration::seconds(10);
-    let result = MOCK_BASIC_APP
+    let result = app
         .send_task(AddTask::new(1, 2).with_expires(dt))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
-    assert!(&message.headers.expires.is_some());
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
+    assert!(message.headers.expires.is_some());
 }
 
 #[tokio::test]
 async fn test_send_task_with_content_type() {
-    let result = MOCK_BASIC_APP
+    let app = build_basic_app().await;
+    let result = app
         .send_task(AddTask::new(1, 2).with_content_type(MessageContentType::Yaml))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
-    assert!(&message.properties.content_type == "application/x-yaml");
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
+    assert!(message.properties.content_type == "application/x-yaml");
 }
 
 #[tokio::test]
 async fn test_send_task_with_time_limit() {
-    let result = MOCK_BASIC_APP
+    let app = build_basic_app().await;
+    let result = app
         .send_task(AddTask::new(1, 2).with_time_limit(5))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
     assert!(message.headers.timelimit == (None, Some(5)));
 }
 
 #[tokio::test]
 async fn test_send_task_with_hard_time_limit() {
-    let result = MOCK_BASIC_APP
+    let app = build_basic_app().await;
+    let result = app
         .send_task(AddTask::new(1, 2).with_hard_time_limit(5))
         .await
         .unwrap();
-    let sent_tasks = MOCK_BASIC_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
     assert!(message.headers.timelimit == (Some(5), None));
 }
 
 #[tokio::test]
 async fn test_configured_app_send_task_app_defaults() {
-    let result = MOCK_CONFIGURED_APP
-        .send_task(AddTask::new(1, 2))
-        .await
-        .unwrap();
-    let sent_tasks = MOCK_CONFIGURED_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
+    let app = build_configured_app().await;
+    let result = app.send_task(AddTask::new(1, 2)).await.unwrap();
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
     assert!(message.headers.timelimit == (None, Some(10)));
-    assert!(&message.properties.content_type == "application/x-yaml");
+    assert!(message.properties.content_type == "application/x-yaml");
 }
 
 #[tokio::test]
 async fn test_configured_app_send_task_task_defaults() {
-    let result = MOCK_CONFIGURED_APP
-        .send_task(MultiplyTask::new(1, 2))
-        .await
-        .unwrap();
-    let sent_tasks = MOCK_CONFIGURED_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
+    let app = build_configured_app().await;
+    let result = app.send_task(MultiplyTask::new(1, 2)).await.unwrap();
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
     assert!(message.headers.timelimit == (Some(10), Some(5)));
-    assert!(&message.properties.content_type == "application/x-yaml");
+    assert!(message.properties.content_type == "application/x-yaml");
 }
 
 #[tokio::test]
 async fn test_configured_app_send_task_request_overrides() {
-    let result = MOCK_CONFIGURED_APP
+    let app = build_configured_app().await;
+    let result = app
         .send_task(
             MultiplyTask::new(1, 2)
                 .with_time_limit(2)
@@ -270,8 +264,8 @@ async fn test_configured_app_send_task_request_overrides() {
         )
         .await
         .unwrap();
-    let sent_tasks = MOCK_CONFIGURED_APP.broker.sent_tasks.read().await;
-    let message = sent_tasks.get(&result.task_id).unwrap();
+    let sent_tasks = app.broker.sent_tasks.read().await;
+    let message = &sent_tasks.get(&result.task_id).unwrap().0;
     assert!(message.headers.timelimit == (Some(10), Some(2)));
-    assert!(&message.properties.content_type == "application/json");
+    assert!(message.properties.content_type == "application/json");
 }