+++ tests/codegen.rs
@@ -1,2 +1,3 @@
 mod app_codegen;
+mod beat_codegen;
 mod task_codegen;