+++ CHANGELOG.md
@@ -7,8 +7,63 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added the `CronSchedule` struct to support Celery's
+  [crontab](https://docs.celeryproject.org/en/stable/reference/celery.schedules.html#celery.schedules.crontab)
+  schedules.
+
 ### Changed
 
+- ⚠️ **BREAKING CHANGE** ⚠️
+
+  To improve the `app!` and `beat!` macros and accomodate custom `Broker`s and `SchedulerBackend`s,
+  we've had to make breaking changes to the way these macros are invoked.
+
+  The biggest change is that the macros now return a future of `Result<Celery>` or `Result<Beat>`.
+  This means you must now call `.await?` on the return value of the macro.
+
+  The other change is that you must now supply the actual `Broker` type.
+  Previously, you could write something like `broker = AMQP { "amqp://my-broker-url" }`,
+  but now you have to write it like `broker = celery::broker::AMQPBroker { "amqp://my-broker-url" }`.
+
+  For a concrete example of these changes, the old way looked like this:
+
+
+  ```rust
+  #[tokio::main]
+  async fn main() -> anyhow::Result<()> {
+      let app = celery::app!(
+          broker = AMQP { "amqp://my-broker-url" },
+          tasks = [add],
+          task_routes = ["*" => "celery"],
+      );
+
+      // ...
+
+      Ok(())
+  }
+  ```
+
+  Whereas now that will look like this:
+
+  ```rust
+  #[tokio::main]
+  async fn main() -> anyhow::Result<()> {
+      let app = celery::app!(
+          broker = celery::broker::AMQPBroker { "amqp://my-broker-url" },
+          tasks = [add],
+          task_routes = ["*" => "celery"],
+      ).await?;
+
+      // ...
+
+      Ok(())
+  }
+  ```
+
+- Celery apps no longer need to have static lifetimes. To remove this constraint, we changed
+  `Celery::consume` to take `&Arc<Self>` instead of a static reference to `self`.
 - Now using `tokio-amqp` internally with `lapin`.
 - Drop explicit dependency on amq-protocol
 
@@ -27,9 +82,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added more tests to the `app` module.
 - Added a `prelude` module.
 - Added support for YAML, MsgPack, and Pickle content types, behind the `extra_content_types` feature flag.
-- Added the `CronSchedule` struct to support Celery's
-  [crontab](https://docs.celeryproject.org/en/stable/reference/celery.schedules.html#celery.schedules.crontab)
-  schedules.
 
 ### Changed